requirements.txt
python-telegram-bot==20.7
yt-dlp>=2024.0.0
instaloader==4.10.3
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
ffmpeg-python==0.2.0
